--- user/src/main/java/com/business/user/application/dto/request/AssignDeliveryDriverRequestDto.java
@@ -1,16 +1,0 @@
-package com.business.user.application.dto.request;
-
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AssignDeliveryDriverRequestDto {
-
-  private UUID deliveryId;
-}
