+++ user/src/main/java/com/business/user/deliverydriver/domain/entity/DriverType.java
@@ -1,0 +1,7 @@
+package com.business.user.deliverydriver.domain.entity;
+
+public enum DriverType {
+
+  HUB,
+  COMPANY
+}
