--- user/src/main/java/com/business/user/domain/entity/DriverType.java
@@ -1,7 +1,0 @@
-package com.business.user.domain.entity;
-
-public enum DriverType {
-
-  HUB,
-  COMPANY
-}
