--- user/src/main/java/com/business/user/presentation/controller/UserController.java
@@ -1,15 +1,0 @@
-package com.business.user.presentation.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/users")
-public class UserController {
-    @GetMapping("/health-check")
-    public ResponseEntity<String> healthCheck() {
-        return ResponseEntity.ok("Health Check OK");
-    }
-}
